cleanup: remove dead code, dedup utilities, expand MetricFormatter
- Delete dead ByteUtils.kt (exact duplicate of core/model CommonUtils.kt)
- Move ByteUtilsTest to core/model as CommonUtilsTest
- Extract normalizeAddress() to core:common AddressUtils.kt, remove
  duplicates from DatabaseConstants.kt and MeshPrefsImpl.kt
- Migrate PowerMetrics.kt (6 calls), SignalMetrics.kt (6 calls), and
  NodeDetailsSection.kt (2 calls) from formatString to MetricFormatter
- Net: -90 lines, 3 fewer formatString imports
