+++ core/model/src/commonTest/kotlin/org/meshtastic/core/model/util/CommonUtilsTest.kt
@@ -14,12 +14,12 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package org.meshtastic.core.common
+package org.meshtastic.core.model.util
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-class ByteUtilsTest {
+class CommonUtilsTest {
 
     @Test
     fun testByteArrayOfInts() {
